requirements/apps.txt
-e svn+http://code.djangoproject.com/svn/django/branches/releases/1.3.X#egg=django
fabric

# requirements for h1ds core
south
django-celery
-e git+git://code.h1svr.anu.edu.au/h1ds/django-openid-auth.git#egg=django_openid_auth

-e git+git://code.h1svr.anu.edu.au/h1ds/h1ds_core.git#egg=h1ds_core


